Back to int.

+++ src/test/java/com/nickbenn/adventofcode/day11/MonkeyInTheMiddleTest.java
@@ -3,9 +3,7 @@ package com.nickbenn.adventofcode.day11;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
-import java.math.BigInteger;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 
@@ -20,7 +18,7 @@ class MonkeyInTheMiddleTest {
 
   @ParameterizedTest
   @CsvFileSource(resources = "cases.csv", numLinesToSkip = 1)
-  void getTopProduct(int part, int rounds, BigInteger divisor, long expected) {
+  void getTopProduct(int part, int rounds, int divisor, long expected) {
     assertEquals(expected, mitm.getTopProduct(rounds, divisor));
   }
 }
